Prepare for 2.19 release
Change the docker/httpd-2.4/Dockerfile to use apt-get to install
apache and mod_perl instead of building perl from source.

Add README-docker that describes how to use Docker to test this distro.

Fix a COPY directive in docker/httpd-2.2/Dockerfile to work when
invoked from the top level of this distro.

Update README to mention the docker stuff and use metacpan for
finding this distro.